Get list of PEcAn Runs that belong to a specific workflow

Retrive a list of all PEcAn runs from the database that belong to a specific workflow

get.runs(server, workflow_id, offset = 0, limit = 50)

Arguments

server

Server object obtained using the connect() function

workflow_id

ID of the PEcAn workflow for which all runs are to be retrieved

offset

The number of workflows to skip before starting to collect the result set.

limit

The number of workflows to return (Available values : 10, 20, 50, 100, 500)

Value

Response obtained from the /api/runs/ endpoint with relevant query parameters

Author

Tezan Sahu

Examples

server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")

# Get details of all runs for the workflow with ID '1000009172'
res <- get.runs(server, workflow_id=1000009172)
#> Sorry! Server not responding